requirements.txt
asgiref==3.2.10
Django==3.1
django-appconf==1.0.4
django-classy-tags==2.0.0
django-cors-headers==3.5.0
django-easy-maps==1.1.2
django-filter==2.3.0
django-phone-field==1.8.1
djangorestframework==3.12.2
geographiclib==1.50
geopy==2.1.0
importlib-metadata==3.1.1
Markdown==3.3.3
mock==4.0.3
mysqlclient==2.0.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
zipp==3.4.0
